% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_popset.R
\name{create_popset}
\alias{create_popset}
\title{Process raw Endnote ris file with PubMed data into population norm set}
\usage{
create_popset(risfile, vocabulary = "MeSH")
}
\arguments{
\item{risfile}{a RIS file produced with Endnote}

\item{vocabulary}{the name of the controlled vocabularay, usually "MeSH"}
}
\value{
a list of objects
}
\description{
Process raw Endnote ris file with PubMed data into population norm set
}
\examples{
# import a risfile
\dontrun{
create_popset(expected.ris)
}
}
