% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_z_scores.R
\name{print_z_scores}
\alias{print_z_scores}
\title{Z-Score tables with German column names ready for exporting}
\usage{
print_z_scores(z_scores_object, terms)
}
\arguments{
\item{z_scores_object}{an object created with z_scores()}

\item{terms}{the terms to be exported, can be either "freetext", "MeSH" or "qualifier"}
}
\value{
a data frame
}
\description{
Z-Score tables with German column names ready for exporting
}
\examples{
path <- system.file("extdata", "example_ris.txt", package = "searchterms", mustWork = T)
data <- z_scores(path)
print_z_scores(data, terms = "freetext")
}
