% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{z_scores}
\alias{z_scores}
\title{Z score tables}
\usage{
z_scores(
  risfile,
  risfile_population,
  load_popset = TRUE,
  dev_set = FALSE,
  seed = NULL
)
}
\arguments{
\item{risfile}{a data frame containing a testset}

\item{risfile_population}{a dataframe containing a population set}

\item{load_popset}{logical, should internal population set be applied (default) or should a customized population set be calculated}

\item{dev_set}{logical, if testset should be randomly divided into a development and validation set select TRUE?}

\item{seed}{should be included, to reproduce a created development set from a previous session}
}
\value{
a list
}
\description{
Z score tables
}
\examples{
path <- system.file("extdata", "example_ris.txt", package = "searchterms")
z_scores(path)

}
